Tracing layer: allow use of cached node list entries for zooming
If a cached entry doesn't match exactly, but has a bounding box that
encloses the one of the current request and if it didn't drop any nodes
(by reaching the maximum nodes limit), the cached entry will now be used
as well. To do this, the LRUCache now allows to iterate all its items
without marking them as accessed. This makes zooming in (and back out)
faster, but causes more elements to being rendered (because *all* nodes
from a larger bounding box might be processed). To mitigate this
somewhat, the smallest cache entry that can be found is used and from
nodes on the current section only those are allowed that are within the
XY bounds of the request bounding box.